add a new app 'Money'
this app will be used to track expenses
