+++ admin_site/admin.py
@@ -3,21 +3,21 @@ from django.contrib import admin
 from django.contrib.auth.admin import UserAdmin, GroupAdmin
 from django.contrib.auth.models import User, Group
 
-from portfolio import models, finance_util
+from portfolio import models as portfolio_models, finance_util
+from money import models as money_models
 
-
-class PortfolioAdminSite(admin.AdminSite):
+class HomieAdminSite(admin.AdminSite):
     site_title = settings.MODE.get_site_title_for('Homie')
     site_header = 'Homie'
     index_title = ''
 
 
-portfolio_admin_site = PortfolioAdminSite(name='portfolio_admin')
-portfolio_admin_site.register(User, UserAdmin)
-portfolio_admin_site.register(Group, GroupAdmin)
+homie_admin_site = HomieAdminSite(name='portfolio_admin')
+homie_admin_site.register(User, UserAdmin)
+homie_admin_site.register(Group, GroupAdmin)
 
 
-@admin.register(models.Transaction, site=portfolio_admin_site)
+@admin.register(portfolio_models.Transaction, site=homie_admin_site)
 class TransactionAdmin(admin.ModelAdmin):
     list_display = ('symbol', 'price', 'type')
     list_filter = ('type',)
@@ -34,12 +34,12 @@ class TransactionAdmin(admin.ModelAdmin):
 
 
 class TransactionInline(admin.TabularInline):
-    model = models.Transaction
+    model = portfolio_models.Transaction
     fields = ('price', 'fee', 'subtotal', 'transaction_time')
     readonly_fields = ('subtotal',)
 
 
-@admin.register(models.Holding, site=portfolio_admin_site)
+@admin.register(portfolio_models.Holding, site=homie_admin_site)
 class HoldingAdmin(admin.ModelAdmin):
     list_display = ('symbol', 'amount', 'total_value')
     ordering = ('symbol', )
@@ -63,6 +63,11 @@ class HoldingAdmin(admin.ModelAdmin):
             **(extra_context or {}),
             'net_worth': finance_util.get_net_worth()
         })
+
+
+@admin.register(money_models.Wallet, site=homie_admin_site)
+class WalletAdmin(admin.ModelAdmin):
+    readonly_fields = ('create_time', 'update_time')
