+++ admin_site/management/commands/init_db.py
@@ -1,6 +1,7 @@
 from django.core.management.base import BaseCommand
 from faker import Faker
 
+from money.models import Wallet
 from portfolio import time_util
 from portfolio.const import TransactionType, DEPOSIT_SYMBOL
 from portfolio.models import Transaction, Holding
@@ -13,6 +14,10 @@ class Command(BaseCommand):
     Faker.seed(0)
 
     def handle(self, *args, **options):
+        self.init_portfolio()
+        self.init_money()
+
+    def init_portfolio(self):
         # add a few deposits
         Transaction.objects.get_or_create(
             symbol=DEPOSIT_SYMBOL,
@@ -26,7 +31,6 @@ class Command(BaseCommand):
             amount=1,
             transaction_time=time_util.now().subtract(self.fake.pyint(-30, 0)),
         )
-
         buys = [
             Transaction.objects.get_or_create(
                 symbol=self.fake.tld().upper(),
@@ -37,7 +41,6 @@ class Command(BaseCommand):
             )[0]
             for _ in range(0, 10)
         ]
-
         for buy in buys:
             for i in range(0, self.fake.random.randint(0, 2)):
                 holding = Holding.objects.get(symbol=buy.symbol)
@@ -55,3 +58,7 @@ class Command(BaseCommand):
                             buy.transaction_time.add(days=self.fake.pyint(0, 10))
                         ),
                     )
+
+    def init_money(self):
+        Wallet.objects.get_or_create(name="Bank")
+        Wallet.objects.get_or_create(name="Credit Card")
