+++ homie/urls.py
@@ -1,10 +1,10 @@
 from django.urls import include, path
 from django.views.generic import RedirectView
 
-from portfolio.admin import portfolio_admin_site
+from admin_site.admin import homie_admin_site
 
 urlpatterns = [
-    path('admin/', portfolio_admin_site.urls),
+    path('admin/', homie_admin_site.urls),
     path('', RedirectView.as_view(url='admin', permanent=True)),
     path('portfolio', include('portfolio.urls'))
 ]
