+++ money/apps.py
@@ -1,8 +1,6 @@
 from django.apps import AppConfig
 
 
-class PortfolioConfig(AppConfig):
-    name = 'portfolio'
+class MoneyConfig(AppConfig):
+    name = 'money'
     default_auto_field = 'django.db.models.BigAutoField'
-
-
