+++ portfolio/tests/test_init_db.py
@@ -1,8 +1,7 @@
 from django.core.management import call_command
 from django.test import TestCase
 
-from portfolio.const import TransactionType
-from portfolio.management.commands import init_db
+from admin_site.management.commands import init_db
 from portfolio.models import Transaction, Holding
 
 
@@ -13,7 +12,6 @@ class InitDBTests(TestCase):
         call_command(init_db.Command())
 
     def test_sell_txn(self):
-        transactions = Transaction.objects.order_by('transaction_time').all()
         holdings = Holding.objects.all()
         for holding in holdings:
             self.assertGreaterEqual(holding.amount, 0)
